// src/modules/responsive-image/types.tsx
export interface IIMAGE_DIMENSIONS {
    BACKGROUND_IMAGE: string; //        16:9    1920
    LIGHT_BOX_IMAGES: string; //         16:5    1600
    HERO_IMAGE: string; //              16:9    1280
    BLOG_IMAGE: string; //              3:2     1200
    PROFILE_IMAGE: string; //           1:1     800
    HERO_IMAGE_SQUARE: string; //       1:1     700
    LIGHT_BOX_IMAGES_SQUARE: string; //  1:1     500
    WEBSITE_BANNER: string; //          1:1     250
    LOGO_RECTANGLE: string; //          3:2     250
    THUMBNAIL_IMAGE: string; //         1:1     150
    LOGO_SQUARE: string; //             2:3     100
    SOCIAL_MEDIA_ICONS: string; //      1:1     32
    FAVICON: string; //                 1:1     16
}